ir: move printing logic into `hash-ir-utils`
