+++ compiler/hash-ir/Cargo.toml
@@ -9,7 +9,6 @@ doctest = false
 
 [dependencies]
 fixedbitset = "0.4.2"
-html-escape = "0.2.12"
 
 hash-ast = { path = "../hash-ast" }
 hash-attrs = { path = "../hash-attrs" }
