+++ compiler/hash-lower/Cargo.toml
@@ -15,6 +15,7 @@ hash-ast = { path = "../hash-ast" }
 hash-attrs = { path = "../hash-attrs" }
 hash-intrinsics = { path = "../hash-intrinsics" }
 hash-ir = { path = "../hash-ir" }
+hash-ir-utils = { path = "../hash-ir-utils" }
 hash-layout = { path = "../hash-layout" }
 hash-pipeline = { path = "../hash-pipeline" }
 hash-reporting = { path = "../hash-reporting" }
